Update index.yml
The metadata tag ms.service is supposed to denote the service slug for the Azure service it covers. The slug `cloud-services` is used for the service "Azure Cloud Services". I don't think this content is related to Cloud Services, so I am removing the tag. Having this tag in this content is skewing our doc metrics.